requirements-dev.txt
python-frontmatter==1.1.0
pyyaml==6.0.2
pytest==8.3.4
